+++ src/Migrate/test/New-AzMigrateProject.Tests.ps1
@@ -12,7 +12,7 @@ while(-not $mockingPath) {
 . ($mockingPath | Select-Object -First 1).FullName
 
 Describe 'New-AzMigrateProject' {
-    It 'PutExpandedCustom' -skip {
+    It 'PutExpandedCustom' -Skip {
         $projName = "AzMigratePwshTestProj123"
         $project = New-AzMigrateProject -Name $projName -ResourceGroupName $env.migResourceGroup -SubscriptionId $env.migSubscriptionId -Location "centralus"
         $project.Name | Should -Be $projName