+++ src/Resources/Resources/ChangeLog.md
@@ -20,6 +20,9 @@
 
 ## Upcoming Release
 
+## Version 3.2.0
+* Added support for -QueryString parameter in New-Az*Deployments cmdlets
+
 ## Version 3.1.1
 * Fixed a NullRef exception issue in `New-AzureManagedApplication` and `Set-AzureManagedApplication`.
 * Updated Azure Resource Manager SDK to use latest DeploymentScripts GA api-version: 2020-10-01.
@@ -34,6 +37,8 @@
 * Added culture info to parse methods
 * Fixed issue where attempts to deploy template specs from a subscription outside of the current subscription context would fail
 * Changed Double parser for version parser
+* Changed New-AzRoleAssignment to include principal type during calls
+* Changed Set-AzRoleAssignment to include principal type during calls
 
 ## Version 3.0.0
 * Fixed parsing bug